:recycle:実装をきれいにした

+++ class.go
@@ -2,20 +2,20 @@ package main
 
 // Class はツクールMVにおける職業
 type Class struct {
-	ExpParams []float32 `json:"expParams"`
-	ID        float32   `json:"id"`
+	ExpParams []float64 `json:"expParams"`
+	ID        float64   `json:"id"`
 	Learnings []struct {
-		Level   float32 `json:"level"`
+		Level   float64 `json:"level"`
 		Note    string  `json:"note"`
-		SkillID float32 `json:"skillId"`
+		SkillID float64 `json:"skillId"`
 	} `json:"learnings"`
 	Name   string      `json:"name"`
 	Note   string      `json:"note"`
-	Params [][]float32 `json:"params"`
+	Params [][]float64 `json:"params"`
 	Traits []struct {
-		Code   float32 `json:"code"`
-		DataID float32 `json:"dataId"`
-		Value  float32 `json:"value"`
+		Code   float64 `json:"code"`
+		DataID float64 `json:"dataId"`
+		Value  float64 `json:"value"`
 	} `json:"traits"`
 }
